requirements.txt
polars
altair
pygsheets
streamlit_cookies_controller
st_pages